Fix last memleaks spotted by coverity
